requirements.txt
streamlit>=1.45.0
gspread>=6.0.0
pandas>=2.0.0
pyperclip>=1.8.2
google-auth>=2.0.0
